pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ebook-scan"
version = "1.0.0"
description = "A tool to scan e-books and convert them to PDF"
readme = "README.md"
authors = [
    {name = "EastShine", email = "eastshine@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "PySide6>=6.0.0",
    "mss>=7.0.0",
    "pynput>=1.7.6",
    "Pillow>=10.0.0",
    "PyAutoGUI>=0.9.54",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
ebook-scan = "ebook_scan.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/ebook-scan-py"
Repository = "https://github.com/yourusername/ebook-scan-py"
Issues = "https://github.com/yourusername/ebook-scan-py/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88